Rename AnnoDesigner.viewmodel -> AnnoDesigner.ViewModels
