+++ AnnoDesigner/StatisticsView.xaml
@@ -4,7 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:local="clr-namespace:AnnoDesigner"
-             xmlns:viewModel="clr-namespace:AnnoDesigner.viewmodel"
+             xmlns:viewModel="clr-namespace:AnnoDesigner.ViewModels"
              xmlns:converter="clr-namespace:AnnoDesigner.Core.Converter;assembly=AnnoDesigner.Core"
              mc:Ignorable="d"
              d:DataContext="{d:DesignInstance Type=viewModel:StatisticsViewModel, IsDesignTimeCreatable=False}"
