+++ AnnoDesigner/ViewModels/StatisticsViewModel.cs
@@ -10,7 +10,7 @@ using AnnoDesigner.Core.Models;
 using AnnoDesigner.Core.Presets.Models;
 using AnnoDesigner.Models;
 
-namespace AnnoDesigner.viewmodel
+namespace AnnoDesigner.ViewModels
 {
     public class StatisticsViewModel : Notify
     {
