+++ AnnoDesigner/ViewModels/WelcomeViewModel.cs
@@ -9,7 +9,7 @@ using System.Windows.Input;
 using AnnoDesigner.Core.Models;
 using AnnoDesigner.Models;
 
-namespace AnnoDesigner.viewmodel
+namespace AnnoDesigner.ViewModels
 {
     public class WelcomeViewModel : Notify
     {
