+++ Tests/AnnoDesigner.Tests/MainViewModelTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using AnnoDesigner.viewmodel;
+using AnnoDesigner.ViewModels;
 using Xunit;
 using Moq;
 using AnnoDesigner.Core.Models;
